Move slice to internals
